Refactoring, combining the separate service and handler into CORSSupport; moving the req and resp handler interfaces and SE impls into their own files
