+++ cors/src/main/java/io/helidon/cors/CrossOriginHelperInternal.java
@@ -87,106 +87,6 @@ public class CrossOriginHelperInternal {
     }
 
     /**
-     * Minimal abstraction of an HTTP request.
-     *
-     * @param <T> type of the request wrapped by the adapter
-     */
-    public interface RequestAdapter<T> {
-
-        /**
-         *
-         * @return possibly unnormalized path from the request
-         */
-        String path();
-
-        /**
-         * Retrieves the first value for the specified header as a String.
-         *
-         * @param key header name to retrieve
-         * @return the first header value for the key
-         */
-        Optional<String> firstHeader(String key);
-
-        /**
-         * Reports whether the specified header exists.
-         *
-         * @param key header name to check for
-         * @return whether the header exists among the request's headers
-         */
-        boolean headerContainsKey(String key);
-
-        /**
-         * Retrieves all header values for a given key as Strings.
-         *
-         * @param key header name to retrieve
-         * @return header values for the header; empty list if none
-         */
-        List<String> allHeaders(String key);
-
-        /**
-         * Reports the method name for the request.
-         *
-         * @return the method name
-         */
-        String method();
-
-        /**
-         * Returns the request this adapter wraps.
-         *
-         * @return the request
-         */
-        T request();
-    }
-
-    /**
-     * Minimal abstraction of an HTTP response.
-     *
-     * <p>
-     * Note to implementers: In some use cases, the CORS support code will invoke the {@code header} methods but not {@code ok}
-     * or {@code forbidden}. See to it that header values set on the adapter via the {@code header} methods are propagated to the
-     * actual response.
-     * </p>
-     *
-     * @param <T> the type of the response wrapped by the adapter
-     */
-    public interface ResponseAdapter<T> {
-
-        /**
-         * Arranges to add the specified header and value to the eventual response.
-         *
-         * @param key header name to add
-         * @param value header value to add
-         * @return the adapter
-         */
-        ResponseAdapter<T> header(String key, String value);
-
-        /**
-         * Arranges to add the specified header and value to the eventual response.
-         *
-         * @param key header name to add
-         * @param value header value to add
-         * @return the adapter
-         */
-        ResponseAdapter<T> header(String key, Object value);
-
-        /**
-         * Returns a response with the forbidden status and the specified error message, without any headers assigned
-         * using the {@code header} methods.
-         *
-         * @param message error message to use in setting the response status
-         * @return the factory
-         */
-        T forbidden(String message);
-
-        /**
-         * Returns a response with only the headers that were set on this adapter and the status set to OK.
-         *
-         * @return response instance
-         */
-        T ok();
-    }
-
-    /**
      * Processes a request according to the CORS rules, returning an {@code Optional} of the response type if
      * the caller should send the response immediately (such as for a preflight response or an error response to a
      * non-preflight CORS request).
@@ -281,7 +181,7 @@ public class CrossOriginHelperInternal {
                     /*
                      * There has been no rejection of the CORS settings, so prep the response headers.
                      */
-                    prepareCORSResponse(crossOrigin, requestAdapter, responseAdapter);
+                    addCORSHeadersToResponse(crossOrigin, requestAdapter, responseAdapter);
                 }
                 return corsResponse;
 
@@ -308,20 +208,29 @@ public class CrossOriginHelperInternal {
         RequestType requestType = requestType(requestAdapter);
 
         if (requestType == RequestType.CORS) {
-            CrossOriginConfig crossOrigin = lookupCrossOrigin(requestAdapter.path(), crossOriginConfigs, secondaryCrossOriginLookup)
+            CrossOriginConfig crossOrigin = lookupCrossOrigin(
+                            requestAdapter.path(),
+                            crossOriginConfigs,
+                            secondaryCrossOriginLookup)
                     .orElseThrow(() -> new IllegalArgumentException(
                             "Could not locate expected CORS information while preparing response to request " + requestAdapter));
-            prepareCORSResponse(
-                    crossOrigin,
-                    requestAdapter,
-                    responseAdapter);
+            addCORSHeadersToResponse(crossOrigin, requestAdapter, responseAdapter);
         }
     }
 
+    /**
+     * Prepares a response with CORS headers, if the supplied request is in fact a CORS request.
+     *
+     * @param crossOrigin the CORS settings to apply to this request
+     * @param requestAdapter abstraction of a request
+     * @param responseAdapter abstraction of a response
+     * @param <T> type for the {@code Request} managed by the requestAdapter
+     * @param <U> the type for the HTTP response as returned from the responseSetter
+     */
     public static <T, U> void prepareResponse(CrossOriginConfig crossOrigin,
             RequestAdapter<T> requestAdapter,
             ResponseAdapter<U> responseAdapter) {
-        prepareCORSResponse(crossOrigin, requestAdapter, responseAdapter);
+        addCORSHeadersToResponse(crossOrigin, requestAdapter, responseAdapter);
     }
 
     /**
@@ -353,6 +262,7 @@ public class CrossOriginHelperInternal {
      * Validates information about an incoming request as a CORS request and, if anything is wrong with CORS information,
      * returns an {@code Optional} error response reporting the problem.
      *
+     * @param crossOriginConfig the CORS settings to apply to this request
      * @param requestAdapter abstraction of a request
      * @param responseAdapter abstraction of a response
      * @param <T> type for the request wrapped by the requestAdapter
@@ -379,12 +289,13 @@ public class CrossOriginHelperInternal {
     /**
      * Prepares a CORS response by updating the response's headers.
      *
+     * @param crossOrigin the CORS settings to apply to the response
      * @param requestAdapter request adapter
      * @param responseAdapter response adapter
      * @param <T> type for the request wrapped by the requestAdapter
      * @param <U> type for the response wrapper by the responseAdapter
      */
-    static <T, U> void prepareCORSResponse(CrossOriginConfig crossOrigin,
+    static <T, U> void addCORSHeadersToResponse(CrossOriginConfig crossOrigin,
             RequestAdapter<T> requestAdapter,
             ResponseAdapter<U> responseAdapter) {
         // Add Access-Control-Allow-Origin and Access-Control-Allow-Credentials.
@@ -415,6 +326,7 @@ public class CrossOriginHelperInternal {
      * Having determined that we have a pre-flight request, we will always return either a forbidden or a successful response.
      * </p>
      *
+     * @param crossOrigin the CORS settings to apply to this request
      * @param requestAdapter the request adapter
      * @param responseAdapter the response adapter
      * @param <T> type for the request wrapped by the requestAdapter
@@ -558,7 +470,7 @@ public class CrossOriginHelperInternal {
         int length = path.length();
         int beginIndex = path.charAt(0) == '/' ? 1 : 0;
         int endIndex = path.charAt(length - 1) == '/' ? length - 1 : length;
-        return (endIndex<= beginIndex) ? "" : path.substring(beginIndex, endIndex);
+        return (endIndex <= beginIndex) ? "" : path.substring(beginIndex, endIndex);
     }
 
     /**
