+++ cors/src/test/java/io/helidon/cors/TestUtil.java
@@ -31,7 +31,7 @@ import io.helidon.webserver.ServerConfiguration;
 import io.helidon.webserver.WebServer;
 
 import static io.helidon.cors.CORSTestServices.SERVICE_3;
-import static io.helidon.cors.CrossOriginConfig.CORS_CONFIG_KEY;
+import static io.helidon.cors.CORSSupport.CORS_CONFIG_KEY;
 
 public class TestUtil {
 
@@ -61,22 +61,22 @@ public class TestUtil {
         /*
          * Use the default config for the service at "/greet" and then programmatically add the config for /cors3.
          */
-        CrossOriginService.Builder corsSupportBuilder = CrossOriginService.builder();
+        CORSSupport.Builder corsSupportBuilder = CORSSupport.builder();
         corsSupportBuilder.addCrossOrigin(SERVICE_3.path(), cors3COC);
 
         /*
          * Load a specific config for "/othergreet."
          */
         Config twoCORSConfig = minimalConfig(ConfigSources.classpath("twoCORS.yaml"));
-        CrossOriginService.Builder twoCORSSupportBuilder =
-                CrossOriginService.builder().config(twoCORSConfig.get(CORS_CONFIG_KEY));
+        CORSSupport.Builder twoCORSSupportBuilder =
+                CORSSupport.builder().config(twoCORSConfig.get(CORS_CONFIG_KEY));
 
         Routing.Builder builder = Routing.builder()
                 .register(GREETING_PATH,
-                          CrossOriginService.fromConfig(), // use "cors" from default app config
+                          CORSSupport.fromConfig(), // use "cors" from default app config
                           new GreetService())
                 .register(OTHER_GREETING_PATH,
-                          CrossOriginService.fromConfig(twoCORSConfig.get(CORS_CONFIG_KEY)), // custom config - get "cors" yourself
+                          CORSSupport.fromConfig(twoCORSConfig.get(CORS_CONFIG_KEY)), // custom config - get "cors" yourself
                           new GreetService("Other Hello"));
 
         return builder;
